gnn_api/requirements.txt
Flask
tensorflow==2.4.1
dgl
torch
PyJWT==2.0.1
python-dotenv==0.17.0
psycopg2
coverage
pytest
Pillow
scikit-image==0.16.2
scikit-learn==0.22.2.post1